Remove unnecessary Latest release step, clarify
The last GitHub action build step to update the manual 'latest' release and tag is unnecessary as Github keeps a latest release reference automatically *and* didn't work in my testing. Removed because its just confusing, IMO :)

Also added a small clarification in the README about the URL to use to point to the module

Aaaand some typo fixes! Yay!